IMG_libpng.c: fix build error due to -Wsign-compare
(cherry picked from commit f3c2cd623ce125eee77e316b6ba479368d72c0b1)

+++ src/IMG_libpng.c
@@ -1502,7 +1502,7 @@ bool IMG_CreateAPNGAnimationDecoder(IMG_AnimationDecoder *decoder, SDL_Propertie
         }
     }
 
-    if (!ctx->is_apng || ctx->fctl_count == 0 || ctx->actl.num_frames > ctx->fctl_count) {
+    if (!ctx->is_apng || ctx->fctl_count == 0 || ctx->actl.num_frames > (unsigned int)ctx->fctl_count) {
         SDL_SetError("Not an APNG file or not enough frame control chunks found");
         IMG_AnimationDecoderClose_Internal(decoder);
         return false;
